Move background to wallet bg class and create parent bg class
